% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlo_connect.R
\name{rlo_connect}
\alias{rlo_connect}
\title{Function to connect to the current LibreOfficeWriter document}
\usage{
rlo_connect()
}
\description{
Function to connect to the current LibreOfficeWriter document
}

