% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlo_dispatch.R
\name{rlo_dispatch}
\alias{rlo_dispatch}
\title{Function to use the dispatcher to use GUI functions}
\usage{
rlo_dispatch(URL, properties)
}
\description{
Function to use the dispatcher to use GUI functions
}

