% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlo_figure.R
\name{rlo_figure}
\alias{rlo_figure}
\title{Insert an SVG graph into the connected document}
\usage{
rlo_figure(svg_file, caption, aspect = 1, width = 17, height = aspect *
  width)
}
\arguments{
\item{svg_file}{Path to an SVG file}

\item{caption}{The caption to use for the figure}

\item{aspect}{The aspect ratio (height/width) to use}

\item{width}{The width in cm}

\item{height}{The height in cm}
}
\description{
Inserts a figure at the current position of the view cursor.
}

