% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlo_heading.R
\name{rlo_heading}
\alias{rlo_heading}
\title{Function to use the dispatcher to use GUI functions}
\usage{
rlo_heading(heading, level = 2, break_before = FALSE)
}
\description{
Function to use the dispatcher to use GUI functions
}

