% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlo_start.R
\name{rlo_new}
\alias{rlo_new}
\title{Create a new libreoffice writer document}
\usage{
rlo_new(file, title = NULL, dir = ".", template = NULL,
  overwrite = FALSE, sleep_time = 1)
}
\arguments{
\item{file}{The file to create/open}

\item{title}{The title of the document}

\item{dir}{The directory to write the file to}

\item{template}{The template file if to be used}

\item{overwrite}{Should file be overwritten if it already exists?}

\item{sleep_time}{The time to sleep after starting LibreOffice before
trying to connect}
}
\description{
Create a new libreoffice writer document
}

