% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlo_parstyle.R
\name{rlo_parstyle}
\alias{rlo_parstyle}
\title{Apply a paragraph style if it exists}
\usage{
rlo_parstyle(parstyle, warn = TRUE)
}
\arguments{
\item{parstyle}{The name of the paragraph style to be applied}

\item{warn}{Should missing paragraph styles give a warning?}
}
\description{
Apply a paragraph style if it exists
}

