% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlo_pdf.R
\name{rlo_pdf}
\alias{rlo_pdf}
\title{Export to PDF}
\usage{
rlo_pdf(file = NULL, dir = ".", overwrite = FALSE)
}
\arguments{
\item{file}{The file to create/open}

\item{dir}{The directory to write the file to}

\item{overwrite}{Should file be overwritten if it already exists?}
}
\description{
By default, this function takes the file name and substitutes .pdf for .odt.
Alternatively, an output file name can be given
}
