% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlo_quit.R
\name{rlo_quit}
\alias{rlo_quit}
\title{Quit a libroffice instance}
\usage{
rlo_quit(save = TRUE)
}
\arguments{
\item{save}{Should changes in the current file be saved on closing?}
}
\description{
Quit a libroffice instance
}
