% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlo_text.R
\name{rlo_text}
\alias{rlo_text}
\title{Insert text}
\usage{
rlo_text(text, format = NULL, break_after = FALSE)
}
\arguments{
\item{text}{The text to be inserted}

\item{format}{Format to be used}

\item{break_after}{Should a paragraph break be inserted after the text?}
}
\description{
Insert text
}
