% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlo_dispatch.R
\name{rlo_dispatch}
\alias{rlo_dispatch}
\title{Function to use the dispatcher to use GUI functions}
\usage{
rlo_dispatch(URL, properties)
}
\arguments{
\item{URL}{The URL of the function to be dispatched}

\item{properties}{A list of properties, often difficult to specify because
of missing documentation}
}
\description{
Function to use the dispatcher to use GUI functions
}

