% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlo_dispatch.R
\name{rlo_dispatch}
\alias{rlo_dispatch}
\title{Use the dispatcher for GUI functions}
\usage{
rlo_dispatch(URL, properties = NULL)
}
\arguments{
\item{URL}{The URL of the function to be dispatched}

\item{properties}{If not NULL, a list of properties to specify the call.
This is often difficult to specify because of missing documentation}
}
\description{
Use the dispatcher for GUI functions
}
\examples{
\dontrun{rlo_dispatch(".uno:Save")}
}
