% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlo_heading.R
\name{rlo_heading}
\alias{rlo_heading}
\title{Function to insert a heading}
\usage{
rlo_heading(heading, level = 2, break_before = FALSE)
}
\arguments{
\item{heading}{The text of the heading}

\item{level}{Heading level from 1 to 10}

\item{break_before}{Should a page break before be associated with the heading?}
}
\description{
Function to insert a heading
}

