% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rlo_start.R
\name{rlo_start}
\alias{rlo_start}
\title{Function to start a new jrwb report}
\usage{
rlo_start(file = NULL, title = NULL, template = NULL, overwrite = FALSE,
  sleep_time = 3)
}
\arguments{
\item{file}{The file to create}

\item{title}{The title of the document}

\item{template}{The template file if to be used}

\item{overwrite}{Should an existing file be overwritten?}

\item{sleep_time}{The time to sleep after starting LibreOffice before
trying to connect}
}
\description{
Function to start a new jrwb report
}

