% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlo_start.R
\name{rlo_start}
\alias{rlo_start}
\title{Start libreoffice writer}
\usage{
rlo_start(file = NULL, title = NULL, dir = ".", template = NULL,
  open = !is.null(file), overwrite = FALSE, sleep_time = 1)
}
\arguments{
\item{file}{The file to create/open}

\item{title}{The title of the document}

\item{dir}{The directory to write the file to}

\item{template}{The template file if to be used}

\item{open}{Should file be opened if it already exists?}

\item{overwrite}{Should file be overwritten if it already exists?}

\item{sleep_time}{The time to sleep after starting LibreOffice before
trying to connect}
}
\description{
Start libreoffice writer
}

